+++ roqoqo-quest/src/interface/pragma_operations.rs
@@ -34,6 +34,10 @@ pub fn execute_pragma_repeated_measurement(
             msg: format!("Probabilites from quantum register {:?}", err),
         })?;
     let mut rng = thread_rng();
+    let existing_register = bit_registers
+        .get(operation.readout())
+        .map(|x| x.to_owned())
+        .unwrap_or_else(|| vec![false; usize::try_from(number_qubits).unwrap()]);
     let output_register: &mut BitOutputRegister = bit_registers_output
         .get_mut(operation.readout())
         .ok_or(RoqoqoBackendError::GenericError {
@@ -47,20 +51,27 @@ pub fn execute_pragma_repeated_measurement(
         None => {
             for _ in 0..*operation.number_measurements() {
                 let index = distribution.sample(&mut rng);
-                output_register.push(index_to_qubits(index, number_qubits))
+                let mut tmp = existing_register.clone();
+                for (a, b) in index_to_qubits(index, number_qubits)
+                    .into_iter()
+                    .enumerate()
+                {
+                    tmp[a] = b
+                }
+                output_register.push(tmp)
             }
         }
         Some(mapping) => {
             for _ in 0..*operation.number_measurements() {
                 let index = distribution.sample(&mut rng);
                 let tmp_output = index_to_qubits(index, number_qubits);
-                let mut new_output: Vec<bool> = vec![false; number_qubits as usize];
-                for (k, val) in tmp_output.iter().enumerate() {
+                let mut new_output: Vec<bool> = existing_register.clone();
+                for (k, val) in tmp_output.into_iter().enumerate() {
                     let tmp_index = match mapping.get(&k) {
                         Some(ind) => ind,
                         None => &k,
                     };
-                    new_output[*tmp_index] = *val;
+                    new_output[*tmp_index] = val;
                 }
                 output_register.push(new_output);
             }
@@ -68,6 +79,54 @@ pub fn execute_pragma_repeated_measurement(
     }
     Ok(())
 }
+
+// pub fn execute_measure_qubit_repeated(
+//     operation: &MeasureQubit,
+//     qureg: &mut Qureg,
+//     number_measurements: usize,
+//     bit_registers: &mut HashMap<String, BitRegister>,
+//     bit_registers_output: &mut HashMap<String, BitOutputRegister>,
+// ) -> Result<(), RoqoqoBackendError> {
+//     let mapping: HashMap<usize, usize> = [(*operation.qubit(), *operation.readout_index())]
+//         .into_iter()
+//         .collect();
+//     let number_qubits = qureg.number_qubits();
+//     let probabilities = qureg.probabilites();
+//     let distribution =
+//         WeightedIndex::new(&probabilities).map_err(|err| RoqoqoBackendError::GenericError {
+//             msg: format!("Probabilites from quantum register {:?}", err),
+//         })?;
+//     let mut rng = thread_rng();
+//     let existing_register = bit_registers
+//         .get(operation.readout())
+//         .map(|x| x.to_owned())
+//         .unwrap_or_else(|| vec![false; usize::try_from(number_qubits).unwrap()]);
+//     let output_register: &mut BitOutputRegister = bit_registers_output
+//         .get_mut(operation.readout())
+//         .ok_or(RoqoqoBackendError::GenericError {
+//             msg: format!(
+//                 "Trying to write readout to non-existent register {}",
+//                 operation.readout()
+//             ),
+//         })?;
+//     bit_registers.remove(operation.readout());
+
+//     for _ in 0..number_measurements {
+//         let index = distribution.sample(&mut rng);
+//         let tmp_output = index_to_qubits(index, number_qubits);
+//         let mut new_output: Vec<bool> = existing_register.clone();
+//         for (k, val) in tmp_output.into_iter().enumerate() {
+//             let tmp_index = match mapping.get(&k) {
+//                 Some(ind) => ind,
+//                 None => &k,
+//             };
+//             new_output[*tmp_index] = val;
+//         }
+//         output_register.push(new_output);
+//     }
+
+//     Ok(())
+// }
 
 pub fn execute_pragma_set_state_vector(
     operation: &PragmaSetStateVector,
@@ -149,9 +208,19 @@ pub fn execute_pragma_set_density_matrix(
         // // Variant for column major order (transpose ndarray default row major)
         let mut reals: Vec<f64> = density_matrix.t().iter().map(|x| x.re).collect();
         let mut imags: Vec<f64> = density_matrix.t().iter().map(|x| x.im).collect();
-
+        let start_row: ::std::os::raw::c_longlong = 0;
+        let start_column: ::std::os::raw::c_longlong = 0;
+        let number_amplitudes: ::std::os::raw::c_longlong =
+            imags.len() as ::std::os::raw::c_longlong;
         unsafe {
-            quest_sys::initStateFromAmps(qureg.quest_qureg, reals.as_mut_ptr(), imags.as_mut_ptr())
+            quest_sys::setDensityAmps(
+                qureg.quest_qureg,
+                start_row,
+                start_column,
+                reals.as_mut_ptr(),
+                imags.as_mut_ptr(),
+                number_amplitudes,
+            )
         }
         Ok(())
     } else {
@@ -159,6 +228,31 @@ pub fn execute_pragma_set_density_matrix(
             msg: "Density matrix can not be set on state vector quantum register".to_string(),
         })
     }
+}
+
+pub fn execute_pragma_input_bit(
+    operation: &InputBit,
+    bit_registers: &mut HashMap<String, BitRegister>,
+) -> Result<(), RoqoqoBackendError> {
+    let existing_register: &mut BitRegister =
+        bit_registers
+            .get_mut(operation.name())
+            .ok_or(RoqoqoBackendError::GenericError {
+                msg: format!(
+                    "Trying to write readout to non-existent register {}",
+                    operation.name()
+                ),
+            })?;
+    if operation.index() >= &existing_register.len() {
+        return Err(RoqoqoBackendError::GenericError {
+            msg: format!(
+                "Trying to write readout to non-existent index {}",
+                operation.index()
+            ),
+        });
+    }
+    existing_register[*operation.index()] = *operation.value();
+    Ok(())
 }
 
 // pub fn execute_pragma_random_noise(